third_party/WebKit/LayoutTests/http/tests/devtools/profiler/cpu-profiler-stopped-removed-race-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that profile removal right after stop profiling issued works. Bug 476430.

Profiler was enabled.

Running: testProfiling
Suspending targets
Stop recording command issued.
Recording stopped. profile (should be null): null
Resuming targets

Profiler was disabled.
